implement our length function in go
This does not yet compile because we don't have anything named `length`
to recurse to.

We're using `[]string` to represent a list for simplicity.  The only
operations we allow ourselves are comparing `len(list)` to 0 to check if
we're at the end of the list and `list[1:]` to get the rest of the list
after the current element.