Wrap the entire expression with fmt.Println
This is now a complete working example of the y-combinator implemented
in go, translated from scheme.

Because there are types sprinkled throughout the code, it's a bit easier
to see what is going on.  For example, there is one function within the
y-combinator that has the same signature as our length function, so you
can guess that this is the actual thing that is called when we want to
recurse.

It's still pretty unidiomatic go, and difficult to follow, but now that
it's working in a familiar language, we can use our usual tools to start
breaking it down.  For example, you could step through it with a
debugger, add debug logging, or start to refactor.