inline applyToSelf in the only place it is used
It's a pretty simple function, and what it does is more obvious if it is
inlined.

We've now broken a refactoring rule: we've assigned an expression to a
variable (`thing2`) and then referenced it twice, which we can't do in
the lambda calculus.  So this go code is now different than the original
scheme code.  However, the core concept of the y-combinator is still
there.