rename `thing2` to `recurser`
This is intentionally the same as the name of the parameter to the
function because we pass the function to itself.  This is a key part of
the y-combinator: when the closure executes, it has a reference to
itself, which it preserves in a closure that it returns.  This
self-reference allows it to call itself again in the next iteration of
the recursion.